requirements.txt
Django
djangorestframework
djangorestframework-simplejwt
django-cors-headers
dj-database-url
psycopg2-binary
gunicorn
whitenoise
